**What this PR does / why we need it**:

Without this copy of `stream.Buffer` we could get bytes in a dirty state when accessing later on. `TestFrameMarshalJSONConcurrent` demonstrates this. Also added similar tests for catch possible problems with DataResponse and QueryDataResponse marshaling in the future (those are marshalled fine at the moment).

Overall performance is comparable:

```
name                     old time/op    new time/op    delta
FrameToJSON-12             26.2µs ± 2%    26.5µs ± 1%      ~     (p=0.481 n=10+10)
FrameMarshalJSONStd-12     59.5µs ± 4%    59.7µs ± 0%      ~     (p=0.237 n=10+8)

name                     old alloc/op   new alloc/op   delta
FrameToJSON-12             2.96kB ± 0%    7.85kB ± 0%  +165.41%  (p=0.000 n=10+9)
FrameMarshalJSONStd-12     7.86kB ± 0%   12.76kB ± 0%   +62.32%  (p=0.000 n=8+10)

name                     old allocs/op  new allocs/op  delta
FrameToJSON-12                148 ± 0%       149 ± 0%    +0.68%  (p=0.000 n=10+10)
FrameMarshalJSONStd-12        149 ± 0%       150 ± 0%    +0.67%  (p=0.000 n=10+10)
```